right-click menu to spawn modules and search

+++ Source/CanvasScrollbar.h
@@ -60,7 +60,7 @@ public:
    void Render() override;
    void MouseReleased() override;
    bool MouseMoved(float x, float y) override;
-   bool MouseScrolled(int x, int y, float scrollX, float scrollY) override;
+   bool MouseScrolled(float x, float y, float scrollX, float scrollY) override;
 
 private:
    void OnClicked(float x, float y, bool right) override;
