+++ Source/IClickable.h
@@ -55,7 +55,7 @@ public:
    IClickable* GetParent() const { return mParent; }
    void SetParent(IClickable* parent) { mParent = parent; }
    bool NotifyMouseMoved(float x, float y);
-   bool NotifyMouseScrolled(int x, int y, float scrollX, float scrollY);
+   bool NotifyMouseScrolled(float x, float y, float scrollX, float scrollY);
    virtual void MouseReleased() {}
    virtual void GetDimensions(float& width, float& height)
    {
@@ -92,7 +92,7 @@ public:
 protected:
    virtual void OnClicked(float x, float y, bool right) {}
    virtual bool MouseMoved(float x, float y) { return false; }
-   virtual bool MouseScrolled(int x, int y, float scrollX, float scrollY) { return false; }
+   virtual bool MouseScrolled(float x, float y, float scrollX, float scrollY) { return false; }
 
    float mX;
    float mY;
