+++ Source/SamplePlayer.h
@@ -130,7 +130,7 @@ private:
    void OnClicked(float x, float y, bool right) override;
    bool MouseMoved(float x, float y) override;
    void MouseReleased() override;
-   bool MouseScrolled(int x, int y, float scrollX, float scrollY) override;
+   bool MouseScrolled(float x, float y, float scrollX, float scrollY) override;
 
    float mWidth{ 608 };
    float mHeight{ 150 };
